add built_in_tuple_codegen_tests.rs
