+++ crates/swift-bridge-ir/src/codegen/codegen_tests.rs
@@ -31,6 +31,7 @@ mod already_declared_attribute_codegen_tests;
 mod argument_label_codegen_tests;
 mod async_function_codegen_tests;
 mod boxed_fnonce_codegen_tests;
+mod built_in_tuple_codegen_tests;
 mod conditional_compilation_codegen_tests;
 mod extern_rust_function_opaque_rust_type_argument_codegen_tests;
 mod extern_rust_function_opaque_rust_type_return_codegen_tests;
